CvPcb: fix wxString formatting bug. (fixes lp:1497908)
* Fix wxString formatting assertion in Eeschema generic netlist exporter.

+++ eeschema/netlist_exporters/netlist_exporter_generic.cpp
@@ -181,7 +181,7 @@ XNODE* NETLIST_EXPORTER_GENERIC::makeComponents()
             xsheetpath->AddAttribute( sNames, path->PathHumanReadable() );
             xsheetpath->AddAttribute( sTStamps, path->Path() );
 
-            timeStamp.Printf( sTSFmt, comp->GetTimeStamp() );
+            timeStamp.Printf( sTSFmt, (unsigned long)comp->GetTimeStamp() );
             xcomp->AddChild( node( sTStamp, timeStamp ) );
         }
     }
